refactor modifyPathVelocityDetail
