+++ planning/behavior_velocity_intersection_module/src/util.hpp
@@ -108,7 +108,7 @@ bool hasAssociatedTrafficLight(lanelet::ConstLanelet lane);
  * @fn
  * @brief interpolate PathWithLaneId
  */
-std::optional<InterpolatedPathInfo> generateInterpolatedPath(
+std::optional<intersection::InterpolatedPathInfo> generateInterpolatedPath(
   const lanelet::Id lane_id, const std::set<lanelet::Id> & associative_lane_ids,
   const autoware_auto_planning_msgs::msg::PathWithLaneId & input_path, const double ds,
   const rclcpp::Logger logger);
@@ -136,14 +136,15 @@ mergeLaneletsByTopologicalSort(
  * polygon
  */
 std::optional<size_t> getFirstPointInsidePolygonByFootprint(
-  const lanelet::CompoundPolygon3d & polygon, const InterpolatedPathInfo & interpolated_path_info,
+  const lanelet::CompoundPolygon3d & polygon,
+  const intersection::InterpolatedPathInfo & interpolated_path_info,
   const tier4_autoware_utils::LinearRing2d & footprint, const double vehicle_length);
 
 std::optional<std::pair<
   size_t /* the index of interpolated PathPoint*/, size_t /* the index of corresponding Polygon */>>
 getFirstPointInsidePolygonsByFootprint(
   const std::vector<lanelet::CompoundPolygon3d> & polygons,
-  const util::InterpolatedPathInfo & interpolated_path_info,
+  const intersection::InterpolatedPathInfo & interpolated_path_info,
   const tier4_autoware_utils::LinearRing2d & footprint, const double vehicle_length);
 
 std::vector<lanelet::CompoundPolygon3d> getPolygon3dFromLanelets(
